# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de gestion des congrès, soumissions et inscriptions",
        "title": "Congrès API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retourne les informations du profil de l'utilisateur connecté",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Récupérer le profil utilisateur",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/congres-app_backend_internal_models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Supprime définitivement le compte de l'utilisateur connecté ainsi que ses données associées",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Supprimer le compte utilisateur",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Modifie les informations du profil de l'utilisateur connecté",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Mettre à jour le profil utilisateur",
                "parameters": [
                    {
                        "description": "Informations du profil à mettre à jour",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/congres-app_backend_internal_models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/profile/password": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permet à l'utilisateur connecté de modifier son mot de passe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Changer le mot de passe",
                "parameters": [
                    {
                        "description": "Ancien et nouveau mot de passe",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/congres-app_backend_pkg_utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/congres-app_backend_pkg_utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "congres-app_backend_internal_models.Notification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "notifiable_id": {
                    "type": "string"
                },
                "notifiable_type": {
                    "type": "string"
                },
                "read_at": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "congres-app_backend_internal_models.Soumission": {
            "type": "object",
            "properties": {
                "author_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "document_title": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "raison_rejet": {
                    "type": "string"
                },
                "resume": {
                    "type": "string"
                },
                "statut": {
                    "description": "En attente, Approuvée, Rejetée",
                    "type": "string"
                },
                "submission_type": {
                    "description": "Abstract, Poster, Communication",
                    "type": "string"
                },
                "theme": {
                    "type": "string"
                },
                "topics": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/congres-app_backend_internal_models.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "congres-app_backend_internal_models.User": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "adresse": {
                    "type": "string"
                },
                "biographie": {
                    "type": "string"
                },
                "civilite": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nom": {
                    "type": "string"
                },
                "notifications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/congres-app_backend_internal_models.Notification"
                    }
                },
                "organisme": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "sexe": {
                    "type": "string"
                },
                "soumissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/congres-app_backend_internal_models.Soumission"
                    }
                },
                "telephone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "congres-app_backend_pkg_utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "congres-app_backend_pkg_utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_handlers.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "internal_handlers.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "adresse": {
                    "type": "string"
                },
                "biographie": {
                    "type": "string"
                },
                "civilite": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nom": {
                    "type": "string"
                },
                "organisme": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                },
                "profession": {
                    "type": "string"
                },
                "sexe": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}